Separate Python vs. Go tests for easier testing
This is needed to be able to run them on various version of Python and
Go runtimes, when we can only install one of the specific language
versions or runtimes at a time.

Now, `make py-test` will just run the Python tests while `make go-test`
will run just the Go tests, and `make test` will continue to run the
combined set of tests for simple development workflow.